fix sidebar issue
